requirements.txt
selenium==4.1.0
python-guerrillamail==0.2.0

